+++ src/rainfallqc/__init__.py
@@ -4,7 +4,7 @@ __author__ = """Tom Keel"""
 __email__ = "tomkee@ceh.ac.uk"
 __version__ = "0.2.1"
 
-from rainfallqc import utils
+from rainfallqc import core, utils
 from rainfallqc.checks import (
     comparison_checks,
     gauge_checks,
@@ -19,5 +19,6 @@ __all__ = [
     "gauge_checks",
     "neighbourhood_checks",
     "timeseries_checks",
+    "core",
     "utils",
 ]