+++ src/rainfallqc/utils/neighbourhood_utils.py
@@ -8,6 +8,12 @@ import numpy as np
 import polars as pl
 import xarray as xr
 
+from rainfallqc.utils import spatial_utils
+
+STATION_ID_COL = "station_id"
+START_DATETIME_COL = "start_datetime"
+END_DATETIME_COL = "end_datetime"
+
 
 def get_target_neighbour_non_zero_minima(
     data: pl.DataFrame, target_col: str, other_col: str, default_minima: float = 0.1
@@ -121,6 +127,9 @@ def get_ids_of_n_nearest_overlapping_neighbouring_gauges(
     distance_threshold: int | float,
     n_closest: int,
     min_overlap_days: int,
+    station_id_col: str = STATION_ID_COL,
+    start_datetime_col: str = START_DATETIME_COL,
+    end_datetime_col: str = END_DATETIME_COL,
 ) -> set:
     """
     Get gauge IDs of nearest n time-overlapping neighbouring gauges.
@@ -137,6 +146,12 @@ def get_ids_of_n_nearest_overlapping_neighbouring_gauges(
         Number of closest neighbours.
     min_overlap_days :
         Minimum overlap between target and neighbouring gauges
+    station_id_col :
+        Column name for station ID in gauge_network_metadata (default 'station_id')
+    start_datetime_col  :
+        Column name for start datetime in gauge_network_metadata (default 'start_datetime')
+    end_datetime_col  :
+        Column name for end datetime in gauge_network_metadata (default 'end_datetime')
 
     Returns
     -------
@@ -145,11 +160,17 @@ def get_ids_of_n_nearest_overlapping_neighbouring_gauges(
 
     """
     # 1. Compute distances between neighbours and target
-    neighbour_distances_df = compute_km_distances_from_target_id(gauge_network_metadata, target_id=target_id)
+    neighbour_distances_df = compute_km_distances_from_target_id(
+        gauge_network_metadata, target_id=target_id, station_id_col=station_id_col
+    )
 
     # 2. Compute overlapping days between neighbours and target
     neighbour_overlap_days_df = compute_temporal_overlap_days_from_target_id(
-        gauge_network_metadata, target_id=target_id
+        gauge_network_metadata,
+        target_id=target_id,
+        station_id_col=station_id_col,
+        start_datetime_col=start_datetime_col,
+        end_datetime_col=end_datetime_col,
     )
 
     # 3. Subset n_closest based on distance threshold
@@ -163,8 +184,8 @@ def get_ids_of_n_nearest_overlapping_neighbouring_gauges(
     )
 
     # 5. get all ids of neighbouring gauges
-    neighbour_distances_ids = neighbour_distances_df["station_id"].to_list()
-    neighbour_overlap_ids = neighbour_overlap_days_df["station_id"].to_list()
+    neighbour_distances_ids = neighbour_distances_df[station_id_col].to_list()
+    neighbour_overlap_ids = neighbour_overlap_days_df[station_id_col].to_list()
 
     # 6. Select gauge IDs meeting both conditions
     all_neighbour_ids = set(neighbour_distances_ids).intersection(set(neighbour_overlap_ids))
@@ -202,7 +223,13 @@ def compute_temporal_overlap_days(
     return overlap_days
 
 
-def compute_temporal_overlap_days_from_target_id(gauge_network_metadata: pl.DataFrame, target_id: str) -> pl.DataFrame:
+def compute_temporal_overlap_days_from_target_id(
+    gauge_network_metadata: pl.DataFrame,
+    target_id: str,
+    station_id_col: str,
+    start_datetime_col: str,
+    end_datetime_col: str,
+) -> pl.DataFrame:
     """
     Compute overlap in days between target gauges and its neighbours.
 
@@ -214,6 +241,12 @@ def compute_temporal_overlap_days_from_target_id(gauge_network_metadata: pl.Data
         Metadata for gauge network. Each gauge must have 'longitude' and 'latitude'.
     target_id :
         Target gauge to compare against.
+    station_id_col :
+        Column name for station ID in gauge_network_metadata
+    start_datetime_col  :
+        Column name for start datetime in gauge_network_metadata
+    end_datetime_col  :
+        Column name for end datetime in gauge_network_metadata
 
     Returns
     -------
@@ -222,16 +255,16 @@ def compute_temporal_overlap_days_from_target_id(gauge_network_metadata: pl.Data
 
     """
     # 1. Get target station and start and end date
-    target_station = gauge_network_metadata.filter(pl.col("station_id") == target_id)
+    target_station = gauge_network_metadata.filter(pl.col(station_id_col) == target_id)
     start_1, end_1 = (
-        target_station["start_datetime"].item(),
-        target_station["end_datetime"].item(),
+        target_station[start_datetime_col].item(),
+        target_station[end_datetime_col].item(),
     )
 
     # 2. Compute overlap days between target station to other start and end date
     neighbour_overlap_days = {}
     for other_station_id, start_2, end_2 in gauge_network_metadata[
-        ["station_id", "start_datetime", "end_datetime"]
+        [station_id_col, start_datetime_col, end_datetime_col]
     ].rows():
         if target_id == other_station_id:
             continue
@@ -241,7 +274,7 @@ def compute_temporal_overlap_days_from_target_id(gauge_network_metadata: pl.Data
     # 3. Convert to pl.Dataframe
     neighbour_overlap_days_df = pl.DataFrame(
         {
-            "station_id": neighbour_overlap_days.keys(),
+            station_id_col: neighbour_overlap_days.keys(),
             "overlap_days": neighbour_overlap_days.values(),
         }
     )
@@ -272,7 +305,9 @@ def get_neighbours_with_min_overlap_days(
     return neighbour_overlap_days_df.filter(pl.col("overlap_days") >= min_overlap_days)
 
 
-def compute_km_distances_from_target_id(gauge_network_metadata: pl.DataFrame, target_id: str) -> pl.DataFrame:
+def compute_km_distances_from_target_id(
+    gauge_network_metadata: pl.DataFrame, target_id: str, station_id_col: str
+) -> pl.DataFrame:
     """
     Compute kilometre distances between gauges in network and target gauges.
 
@@ -282,6 +317,8 @@ def compute_km_distances_from_target_id(gauge_network_metadata: pl.DataFrame, ta
         Metadata for gauge network. Each gauge must have 'longitude' and 'latitude'.
     target_id :
         Target gauge to compare against.
+    station_id_col :
+        Column name for station ID in gauge_network_metadata
 
     Returns
     -------
@@ -290,7 +327,7 @@ def compute_km_distances_from_target_id(gauge_network_metadata: pl.DataFrame, ta
 
     """
     # 1. Get target station lat and lon
-    target_station = gauge_network_metadata.filter(pl.col("station_id") == target_id)
+    target_station = gauge_network_metadata.filter(pl.col(station_id_col) == target_id)
     target_latlon = (
         target_station["latitude"].item(),
         target_station["longitude"].item(),
@@ -299,7 +336,7 @@ def compute_km_distances_from_target_id(gauge_network_metadata: pl.DataFrame, ta
     # 2. Calculate lat/lon distances from the target gauge
     neighbour_distances = {}
     for other_station_id, other_lat, other_lon in gauge_network_metadata[
-        ["station_id", "latitude", "longitude"]
+        [station_id_col, "latitude", "longitude"]
     ].rows():
         neighbour_distances[other_station_id] = geopy.distance.geodesic(
             target_latlon, (other_lat, other_lon)
@@ -308,7 +345,7 @@ def compute_km_distances_from_target_id(gauge_network_metadata: pl.DataFrame, ta
     # 3. Convert to pl.Dataframe
     neighbour_distances_df = pl.DataFrame(
         {
-            "station_id": neighbour_distances.keys(),
+            station_id_col: neighbour_distances.keys(),
             "distance": neighbour_distances.values(),
         }
     )
@@ -358,31 +395,57 @@ def get_n_closest_neighbours(
         return sorted_close_neighbours.filter(pl.col("distance") <= nth_distance)
 
 
-def get_nearest_etccdi_val_to_gauge(
-    etccdi_data: xr.Dataset, gauge_lat: int | float, gauge_lon: int | float
+def get_nearest_non_nan_etccdi_val_to_gauge(
+    etccdi_data: xr.Dataset,
+    etccdi_name: str,
+    gauge_lat: int | float,
+    gauge_lon: int | float,
+    max_distance_km: int | float = 500,
 ) -> xr.Dataset:
     """
-    Get the value at the nearest ETCCDI grid cell to the gauge coordinates.
+    Get the value at the nearest non-nan ETCCDI grid cell to the gauge coordinates.
 
     Parameters
     ----------
     etccdi_data
         ETCCDI data with given variable to check
+    etccdi_name :
+        ETCCDI variable name to check
     gauge_lat :
         latitude of the rain gauge
     gauge_lon :
         longitude of the rain gauge
+    max_distance_km :
+        Maximum distance in km to search for a non-nan value (default 500 km)
 
     Returns
     -------
     nearby_etccdi_data :
-        ETCCDI data at the nearest grid cell
+        ETCCDI data at the nearest grid cell with non-nan values
 
     """
-    # 1. Get local data
-    nearby_etccdi_data = etccdi_data.sel(
-        lon=gauge_lon,
-        lat=gauge_lat,
-        method="nearest",
+    # 1. Stack lat/lon into a single dimension
+    stacked = etccdi_data.stack(points=("lat", "lon"))
+
+    # 2. Compute haversine distance to each point
+    dists = spatial_utils.haversine(stacked["lon"], stacked["lat"], gauge_lon, gauge_lat)
+
+    # 3. Sort by distance
+    sorted_idx = dists.argsort()
+
+    # Loop through points in ascending distance order
+    for idx in sorted_idx.values:
+        dist_km = dists.isel(points=idx).item()
+        if dist_km <= max_distance_km:
+            if not np.isnan(stacked[etccdi_name].isel(points=idx)).all():
+                sel_lon = stacked["lon"].isel(points=idx).item()
+                sel_lat = stacked["lat"].isel(points=idx).item()
+                return etccdi_data.sel(lon=sel_lon, lat=sel_lat)
+        else:
+            # Because sorted by distance, if we've exceeded max distance, we can break early
+            break
+
+    raise ValueError(
+        f"""No non-NaN point found within {max_distance_km} km of ({gauge_lat}, {gauge_lon}).
+        Assuming EPSG:4326 coordinates."""
     )
-    return nearby_etccdi_data