+++ tests/unit/test_neighbourhood_utils.py
@@ -5,13 +5,14 @@
 import datetime
 
 import polars as pl
+import pytest
 
-from rainfallqc.utils import neighbourhood_utils
+from rainfallqc.utils import data_readers, neighbourhood_utils
 
 
 def test_compute_distance_from_target_id(gdsr_gauge_network):
     result = neighbourhood_utils.compute_km_distances_from_target_id(
-        gauge_network_metadata=gdsr_gauge_network, target_id="DE_00310"
+        gauge_network_metadata=gdsr_gauge_network, target_id="DE_00310", station_id_col="station_id"
     )
     assert round(result.filter(pl.col("station_id") == "DE_02483")["distance"][0], 2) == 13.13
     assert round(result.filter(pl.col("station_id") == "DE_00310")["distance"][0], 2) == 0.0
@@ -19,7 +20,7 @@ def test_compute_distance_from_target_id(gdsr_gauge_network):
 
 def test_get_n_closest_neighbours(gdsr_gauge_network):
     neighbouring_gauges = neighbourhood_utils.compute_km_distances_from_target_id(
-        gauge_network_metadata=gdsr_gauge_network, target_id="DE_00310"
+        gauge_network_metadata=gdsr_gauge_network, target_id="DE_00310", station_id_col="station_id"
     )
     # distance should be in km
     result = neighbourhood_utils.get_n_closest_neighbours(neighbouring_gauges, distance_threshold=50, n_closest=10)
@@ -48,14 +49,24 @@ def test_compute_temporal_overlap_days():
 
 
 def test_compute_temporal_overlap_days_from_target_id(gdsr_gauge_network):
-    result = neighbourhood_utils.compute_temporal_overlap_days_from_target_id(gdsr_gauge_network, target_id="DE_00310")
+    result = neighbourhood_utils.compute_temporal_overlap_days_from_target_id(
+        gdsr_gauge_network,
+        target_id="DE_00310",
+        station_id_col="station_id",
+        start_datetime_col="start_datetime",
+        end_datetime_col="end_datetime",
+    )
     assert result.filter(pl.col("station_id") == "DE_02483")["overlap_days"][0] == 1825
     assert result.filter(pl.col("station_id") == "DE_00389")["overlap_days"][0] == 425
 
 
 def test_get_neighbours_with_min_overlap_days(gdsr_gauge_network):
     neighbour_overlap_days_df = neighbourhood_utils.compute_temporal_overlap_days_from_target_id(
-        gdsr_gauge_network, target_id="DE_00310"
+        gdsr_gauge_network,
+        target_id="DE_00310",
+        station_id_col="station_id",
+        start_datetime_col="start_datetime",
+        end_datetime_col="end_datetime",
     )
     result = neighbourhood_utils.get_neighbours_with_min_overlap_days(neighbour_overlap_days_df, min_overlap_days=1500)
     assert len(result) == 8
@@ -100,3 +111,16 @@ def test_make_rain_not_minima_column_target_or_neighbour(gauge_comparison_data):
         gauge_comparison_data, target_col="gauge1", other_col="gauge2", data_minima=2.0
     )
     assert result["rain_not_minima"].value_counts().filter(pl.col("rain_not_minima") == 1)["count"].item() == 1
+
+
+def test_get_nearest_non_nan_etccdi_val_to_gauge():
+    etccdi_r99p = data_readers.load_etccdi_data(etccdi_var="R99p")
+
+    result = neighbourhood_utils.get_nearest_non_nan_etccdi_val_to_gauge(
+        etccdi_r99p, etccdi_name="R99p", gauge_lat=50.0, gauge_lon=10.0
+    )
+    assert round(float(result["R99p"].max()), 2) == 91.68
+    with pytest.raises(ValueError):
+        neighbourhood_utils.get_nearest_non_nan_etccdi_val_to_gauge(
+            etccdi_r99p, etccdi_name="R99p", gauge_lat=90.0, gauge_lon=10.0
+        )